+++ CHANGELOG.md
@@ -23,6 +23,8 @@
     3. [](#bugfix)
         - Fixed dismissal links alignment for alerts (#1022)
 3. [Wordpress](#wordpress)
+    2. [](#improved)
+        - Display notification for the logged in user when site is offline (#760) 
     3. [](#bugfix)
         - Fixed plugin settings being disabled when theme failed to load
         - Fixed `Cannot use object of type Closure as array` (#1046)